backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.15.2
python-dotenv==1.1.0
pydantic==2.11.3
pydantic-settings==2.9.1
fhir.resources==8.0.0
httpx==0.28.1
python-multipart==0.0.20
